// src/sdk/generated/models/AlertDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AlertDto
 */
export interface AlertDto {
    /**
     * 
     * @type {number}
     * @memberof AlertDto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof AlertDto
     */
    message: string;
    /**
     * 
     * @type {Date}
     * @memberof AlertDto
     */
    createdAt: Date;
    /**
     * 
     * @type {number}
     * @memberof AlertDto
     */
    alertType: number;
    /**
     * 
     * @type {boolean}
     * @memberof AlertDto
     */
    isActive?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof AlertDto
     */
    resolvedAt?: Date | null;
    /**
     * 
     * @type {number}
     * @memberof AlertDto
     */
    batchId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof AlertDto
     */
    zoneId?: number | null;
}

/**
 * Check if a given object implements the AlertDto interface.
 */
export function instanceOfAlertDto(value: object): value is AlertDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('alertType' in value) || value['alertType'] === undefined) return false;
    return true;
}

export function AlertDtoFromJSON(json: any): AlertDto {
    return AlertDtoFromJSONTyped(json, false);
}

export function AlertDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AlertDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'message': json['message'],
        'createdAt': (new Date(json['createdAt'])),
        'alertType': json['alertType'],
        'isActive': json['isActive'] == null ? undefined : json['isActive'],
        'resolvedAt': json['resolvedAt'] == null ? undefined : (new Date(json['resolvedAt'])),
        'batchId': json['batchId'] == null ? undefined : json['batchId'],
        'zoneId': json['zoneId'] == null ? undefined : json['zoneId'],
    };
}

export function AlertDtoToJSON(json: any): AlertDto {
    return AlertDtoToJSONTyped(json, false);
}

export function AlertDtoToJSONTyped(value?: AlertDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'message': value['message'],
        'createdAt': value['createdAt'].toISOString(),
        'alertType': value['alertType'],
        'isActive': value['isActive'],
        'resolvedAt': value['resolvedAt'] == null ? value['resolvedAt'] : value['resolvedAt'].toISOString(),
        'batchId': value['batchId'],
        'zoneId': value['zoneId'],
    };
}