// src/sdk/generated/models/AuditLogDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AuditLogDto
 */
export interface AuditLogDto {
    /**
     * 
     * @type {number}
     * @memberof AuditLogDto
     */
    id?: number;
    /**
     * 
     * @type {Date}
     * @memberof AuditLogDto
     */
    occurredAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof AuditLogDto
     */
    entityType?: string;
    /**
     * 
     * @type {number}
     * @memberof AuditLogDto
     */
    entityId?: number;
    /**
     * 
     * @type {string}
     * @memberof AuditLogDto
     */
    action?: string;
    /**
     * 
     * @type {string}
     * @memberof AuditLogDto
     */
    userId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuditLogDto
     */
    summary?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuditLogDto
     */
    oldValues?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuditLogDto
     */
    newValues?: string | null;
}

/**
 * Check if a given object implements the AuditLogDto interface.
 */
export function instanceOfAuditLogDto(value: object): value is AuditLogDto {
    return true;
}

export function AuditLogDtoFromJSON(json: any): AuditLogDto {
    return AuditLogDtoFromJSONTyped(json, false);
}

export function AuditLogDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuditLogDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'occurredAt': json['occurredAt'] == null ? undefined : (new Date(json['occurredAt'])),
        'entityType': json['entityType'] == null ? undefined : json['entityType'],
        'entityId': json['entityId'] == null ? undefined : json['entityId'],
        'action': json['action'] == null ? undefined : json['action'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'summary': json['summary'] == null ? undefined : json['summary'],
        'oldValues': json['oldValues'] == null ? undefined : json['oldValues'],
        'newValues': json['newValues'] == null ? undefined : json['newValues'],
    };
}

export function AuditLogDtoToJSON(json: any): AuditLogDto {
    return AuditLogDtoToJSONTyped(json, false);
}

export function AuditLogDtoToJSONTyped(value?: AuditLogDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'occurredAt': value['occurredAt'] == null ? value['occurredAt'] : value['occurredAt'].toISOString(),
        'entityType': value['entityType'],
        'entityId': value['entityId'],
        'action': value['action'],
        'userId': value['userId'],
        'summary': value['summary'],
        'oldValues': value['oldValues'],
        'newValues': value['newValues'],
    };
}