// src/sdk/generated/models/BatchCreateDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BatchCreateDto
 */
export interface BatchCreateDto {
    /**
     * 
     * @type {string}
     * @memberof BatchCreateDto
     */
    batchNumber: string;
    /**
     * 
     * @type {number}
     * @memberof BatchCreateDto
     */
    quantity: number;
    /**
     * 
     * @type {Date}
     * @memberof BatchCreateDto
     */
    expireDate: Date;
    /**
     * 
     * @type {Date}
     * @memberof BatchCreateDto
     */
    dateAdded: Date;
    /**
     * 
     * @type {number}
     * @memberof BatchCreateDto
     */
    medicineId: number;
    /**
     * 
     * @type {number}
     * @memberof BatchCreateDto
     */
    zoneId: number;
}

/**
 * Check if a given object implements the BatchCreateDto interface.
 */
export function instanceOfBatchCreateDto(value: object): value is BatchCreateDto {
    if (!('batchNumber' in value) || value['batchNumber'] === undefined) return false;
    if (!('quantity' in value) || value['quantity'] === undefined) return false;
    if (!('expireDate' in value) || value['expireDate'] === undefined) return false;
    if (!('dateAdded' in value) || value['dateAdded'] === undefined) return false;
    if (!('medicineId' in value) || value['medicineId'] === undefined) return false;
    if (!('zoneId' in value) || value['zoneId'] === undefined) return false;
    return true;
}

export function BatchCreateDtoFromJSON(json: any): BatchCreateDto {
    return BatchCreateDtoFromJSONTyped(json, false);
}

export function BatchCreateDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BatchCreateDto {
    if (json == null) {
        return json;
    }
    return {
        
        'batchNumber': json['batchNumber'],
        'quantity': json['quantity'],
        'expireDate': (new Date(json['expireDate'])),
        'dateAdded': (new Date(json['dateAdded'])),
        'medicineId': json['medicineId'],
        'zoneId': json['zoneId'],
    };
}

export function BatchCreateDtoToJSON(json: any): BatchCreateDto {
    return BatchCreateDtoToJSONTyped(json, false);
}

export function BatchCreateDtoToJSONTyped(value?: BatchCreateDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'batchNumber': value['batchNumber'],
        'quantity': value['quantity'],
        'expireDate': value['expireDate'].toISOString(),
        'dateAdded': value['dateAdded'].toISOString(),
        'medicineId': value['medicineId'],
        'zoneId': value['zoneId'],
    };
}