// src/sdk/generated/models/BatchDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BatchDto
 */
export interface BatchDto {
    /**
     * 
     * @type {number}
     * @memberof BatchDto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof BatchDto
     */
    batchNumber: string;
    /**
     * 
     * @type {number}
     * @memberof BatchDto
     */
    quantity: number;
    /**
     * 
     * @type {Date}
     * @memberof BatchDto
     */
    expireDate: Date;
    /**
     * 
     * @type {Date}
     * @memberof BatchDto
     */
    dateAdded: Date;
    /**
     * 
     * @type {number}
     * @memberof BatchDto
     */
    medicineId: number;
    /**
     * 
     * @type {number}
     * @memberof BatchDto
     */
    zoneId: number;
}

/**
 * Check if a given object implements the BatchDto interface.
 */
export function instanceOfBatchDto(value: object): value is BatchDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('batchNumber' in value) || value['batchNumber'] === undefined) return false;
    if (!('quantity' in value) || value['quantity'] === undefined) return false;
    if (!('expireDate' in value) || value['expireDate'] === undefined) return false;
    if (!('dateAdded' in value) || value['dateAdded'] === undefined) return false;
    if (!('medicineId' in value) || value['medicineId'] === undefined) return false;
    if (!('zoneId' in value) || value['zoneId'] === undefined) return false;
    return true;
}

export function BatchDtoFromJSON(json: any): BatchDto {
    return BatchDtoFromJSONTyped(json, false);
}

export function BatchDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BatchDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'batchNumber': json['batchNumber'],
        'quantity': json['quantity'],
        'expireDate': (new Date(json['expireDate'])),
        'dateAdded': (new Date(json['dateAdded'])),
        'medicineId': json['medicineId'],
        'zoneId': json['zoneId'],
    };
}

export function BatchDtoToJSON(json: any): BatchDto {
    return BatchDtoToJSONTyped(json, false);
}

export function BatchDtoToJSONTyped(value?: BatchDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'batchNumber': value['batchNumber'],
        'quantity': value['quantity'],
        'expireDate': value['expireDate'].toISOString(),
        'dateAdded': value['dateAdded'].toISOString(),
        'medicineId': value['medicineId'],
        'zoneId': value['zoneId'],
    };
}