// src/sdk/generated/models/BatchSearchResultDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BatchSearchResultDto
 */
export interface BatchSearchResultDto {
    /**
     * 
     * @type {number}
     * @memberof BatchSearchResultDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof BatchSearchResultDto
     */
    batchNumber?: string;
    /**
     * 
     * @type {number}
     * @memberof BatchSearchResultDto
     */
    medicineId?: number;
    /**
     * 
     * @type {number}
     * @memberof BatchSearchResultDto
     */
    zoneId?: number;
}

/**
 * Check if a given object implements the BatchSearchResultDto interface.
 */
export function instanceOfBatchSearchResultDto(value: object): value is BatchSearchResultDto {
    return true;
}

export function BatchSearchResultDtoFromJSON(json: any): BatchSearchResultDto {
    return BatchSearchResultDtoFromJSONTyped(json, false);
}

export function BatchSearchResultDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BatchSearchResultDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'batchNumber': json['batchNumber'] == null ? undefined : json['batchNumber'],
        'medicineId': json['medicineId'] == null ? undefined : json['medicineId'],
        'zoneId': json['zoneId'] == null ? undefined : json['zoneId'],
    };
}

export function BatchSearchResultDtoToJSON(json: any): BatchSearchResultDto {
    return BatchSearchResultDtoToJSONTyped(json, false);
}

export function BatchSearchResultDtoToJSONTyped(value?: BatchSearchResultDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'batchNumber': value['batchNumber'],
        'medicineId': value['medicineId'],
        'zoneId': value['zoneId'],
    };
}