// src/sdk/generated/models/MedicineCreateDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MedicineCreateDto
 */
export interface MedicineCreateDto {
    /**
     * 
     * @type {string}
     * @memberof MedicineCreateDto
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof MedicineCreateDto
     */
    description?: string | null;
    /**
     * 
     * @type {number}
     * @memberof MedicineCreateDto
     */
    tempMax: number;
    /**
     * 
     * @type {number}
     * @memberof MedicineCreateDto
     */
    tempMin: number;
    /**
     * 
     * @type {number}
     * @memberof MedicineCreateDto
     */
    humidMax: number;
    /**
     * 
     * @type {number}
     * @memberof MedicineCreateDto
     */
    humidMin: number;
}

/**
 * Check if a given object implements the MedicineCreateDto interface.
 */
export function instanceOfMedicineCreateDto(value: object): value is MedicineCreateDto {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('tempMax' in value) || value['tempMax'] === undefined) return false;
    if (!('tempMin' in value) || value['tempMin'] === undefined) return false;
    if (!('humidMax' in value) || value['humidMax'] === undefined) return false;
    if (!('humidMin' in value) || value['humidMin'] === undefined) return false;
    return true;
}

export function MedicineCreateDtoFromJSON(json: any): MedicineCreateDto {
    return MedicineCreateDtoFromJSONTyped(json, false);
}

export function MedicineCreateDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MedicineCreateDto {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'tempMax': json['tempMax'],
        'tempMin': json['tempMin'],
        'humidMax': json['humidMax'],
        'humidMin': json['humidMin'],
    };
}

export function MedicineCreateDtoToJSON(json: any): MedicineCreateDto {
    return MedicineCreateDtoToJSONTyped(json, false);
}

export function MedicineCreateDtoToJSONTyped(value?: MedicineCreateDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'description': value['description'],
        'tempMax': value['tempMax'],
        'tempMin': value['tempMin'],
        'humidMax': value['humidMax'],
        'humidMin': value['humidMin'],
    };
}