// src/sdk/generated/models/MedicineDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MedicineDto
 */
export interface MedicineDto {
    /**
     * 
     * @type {number}
     * @memberof MedicineDto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof MedicineDto
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof MedicineDto
     */
    description?: string | null;
    /**
     * 
     * @type {number}
     * @memberof MedicineDto
     */
    tempMax: number;
    /**
     * 
     * @type {number}
     * @memberof MedicineDto
     */
    tempMin: number;
    /**
     * 
     * @type {number}
     * @memberof MedicineDto
     */
    humidMax: number;
    /**
     * 
     * @type {number}
     * @memberof MedicineDto
     */
    humidMin: number;
}

/**
 * Check if a given object implements the MedicineDto interface.
 */
export function instanceOfMedicineDto(value: object): value is MedicineDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('tempMax' in value) || value['tempMax'] === undefined) return false;
    if (!('tempMin' in value) || value['tempMin'] === undefined) return false;
    if (!('humidMax' in value) || value['humidMax'] === undefined) return false;
    if (!('humidMin' in value) || value['humidMin'] === undefined) return false;
    return true;
}

export function MedicineDtoFromJSON(json: any): MedicineDto {
    return MedicineDtoFromJSONTyped(json, false);
}

export function MedicineDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MedicineDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'tempMax': json['tempMax'],
        'tempMin': json['tempMin'],
        'humidMax': json['humidMax'],
        'humidMin': json['humidMin'],
    };
}

export function MedicineDtoToJSON(json: any): MedicineDto {
    return MedicineDtoToJSONTyped(json, false);
}

export function MedicineDtoToJSONTyped(value?: MedicineDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'tempMax': value['tempMax'],
        'tempMin': value['tempMin'],
        'humidMax': value['humidMax'],
        'humidMin': value['humidMin'],
    };
}