// src/sdk/generated/models/MedicineSearchResultDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WebApi | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MedicineSearchResultDto
 */
export interface MedicineSearchResultDto {
    /**
     * 
     * @type {number}
     * @memberof MedicineSearchResultDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof MedicineSearchResultDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof MedicineSearchResultDto
     */
    description?: string | null;
}

/**
 * Check if a given object implements the MedicineSearchResultDto interface.
 */
export function instanceOfMedicineSearchResultDto(value: object): value is MedicineSearchResultDto {
    return true;
}

export function MedicineSearchResultDtoFromJSON(json: any): MedicineSearchResultDto {
    return MedicineSearchResultDtoFromJSONTyped(json, false);
}

export function MedicineSearchResultDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): MedicineSearchResultDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function MedicineSearchResultDtoToJSON(json: any): MedicineSearchResultDto {
    return MedicineSearchResultDtoToJSONTyped(json, false);
}

export function MedicineSearchResultDtoToJSONTyped(value?: MedicineSearchResultDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
    };
}